refactored to use subclassed api
