+++ test/issues/issue-08.test.ts
@@ -1,11 +1,11 @@
 import { expect, test } from 'vitest';
 import { z } from 'zod';
-import { createFixture } from 'zod-fixture';
+import { Fixture } from 'zod-fixture';
 
 test('creates object with string property of length', () => {
 	const schema = z.object({
 		lastFour: z.string().length(4),
 	});
-	const fixture = createFixture(schema);
+	const fixture = new Fixture().generate(schema);
 	expect(fixture.lastFour).toHaveLength(4);
 });
